requirements.txt
absl-py>=0.12.0
spacy>=3.0.6
tqdm>=4.60.0
